+++ spec/researches/calculateFleschReadingSpec.js
@@ -46,6 +46,13 @@ describe( "A test that uses the German Flesch Reading", function() {
 	} );
 } );
 
+describe( "A test that uses the Italian Flesch Reading", function() {
+	it( "returns a score", function () {
+		var mockPaper = new Paper( "Che bella cosa e 'una sola giornata, n'aria serena doppo na tempesta.", { locale: "it_IT" } );
+		expect( fleschFunction( mockPaper ) ).toBe( 76.4 );
+	} );
+} );
+
 describe( "A test that returns 0 after sentence formatting", function() {
 	it( "returns a score of 0", function() {
 		var mockPaper = new Paper( "()" );
